refactor(*): add comon file to tests and fix restaurant error test
