+++ backend/.env
@@ -1,2 +1,2 @@
 PORT=3001
-DBNAME=test
+DBNAME=test1
