+++ backend/tests/controllers/common.ts
@@ -1,0 +1,23 @@
+const mongoose = require('mongoose')
+require("dotenv").config()
+
+export async function connectDBForTesting() {
+    try {
+      const dbUri = "mongodb://localhost:27017";
+      const dbName = process.env.DBNAME;
+      await mongoose.connect(dbUri, {
+        dbName,
+        autoCreate: true,
+      });
+    } catch (error) {
+      console.log("DB connect error");
+    }
+}
+  
+export async function disconnectDBForTesting() {
+    try {
+      await mongoose.connection.close();
+    } catch (error) {
+      console.log("DB disconnect error");
+    }
+}
