+++ backend/tests/controllers/restaurant/restaurant_edit.steps.ts
@@ -1,27 +1,8 @@
 import { loadFeature, defineFeature } from "jest-cucumber"
 import axios, { AxiosResponse } from 'axios'
 const mongoose = require('mongoose')
-
-export async function connectDBForTesting() {
-    try {
-      const dbUri = "mongodb://localhost:27017";
-      const dbName = "test";
-      await mongoose.connect(dbUri, {
-        dbName,
-        autoCreate: true,
-      });
-    } catch (error) {
-      console.log("DB connect error");
-    }
-}
-  
-export async function disconnectDBForTesting() {
-    try {
-      await mongoose.connection.close();
-    } catch (error) {
-      console.log("DB disconnect error");
-    }
-}
+require("dotenv").config()
+import {connectDBForTesting, disconnectDBForTesting} from '../common'
 
 const Restaurant = require("../../../models/Restaurant")
